arm: tegra: fix warning to enable -Werror
Fix various warnings to enable -Werror option.

Bug 1211919

Reviewed-on: http://git-master/r/354292

+++ arch/arm/mach-tegra/include/mach/dc.h
@@ -280,7 +280,7 @@ struct tegra_dsi_out {
 	u16		dsi_panel_rst_gpio;
 	u16		dsi_panel_bl_en_gpio;
 	u16		dsi_panel_bl_pwm_gpio;
-	u8		chip_id;
+	u16		chip_id;
 	u8		chip_rev;
 	u8		controller_vs;
